refactor(media): host upload adapters in the media layer with server-free primitives
